// src/services/notification/types/enums.ts
export enum EmailType {
  CANDIDATE_BOOKING_CANCELLATION = 'CANDIDATE_BOOKING_CANCELLATION',
  TRAINER_BOOKER_BOOKING_CANCELLATION = 'TRAINER_BOOKER_BOOKING_CANCELLATION',
}

export enum LetterType {
  BOOKING_CANCELLATION,
}

export enum Target {
  GB = 'gb',
  NI = 'ni',
}

export enum TestType {
  ADIHPT = 'adihpt',
  ADIP1 = 'adip1',
  ADIP1DVA = 'adip1dva',
  AMIP1 = 'amip1',
  CAR = 'car',
  ERS = 'ers',
  LGVCPC = 'lgvcpc',
  LGVCPCC = 'lgvcpcc',
  LGVHPT = 'lgvhpt',
  LGVMC = 'lgvmc',
  MOTORCYCLE = 'motorcycle',
  PCVCPC = 'pcvcpc',
  PCVCPCC = 'pcvcpcc',
  PCVHPT = 'pcvhpt',
  PCVMC = 'pcvmc',
  TAXI = 'taxi',
}